+++ R/gads_get_campaigns.R
@@ -2,9 +2,9 @@
 #'
 #' @inheritParams gads_get_report
 #'
-#' @param fields character vector, list of report fields, all report has own fields list, for example \href{https://developers.google.com/google-ads/api/fields/v10/campaign}{see field list of campaigns report}.
+#' @param fields character vector, list of report fields, all report has own fields list, for example \href{https://developers.google.com/google-ads/api/fields/v23/campaign}{see field list of campaigns report}.
 #'
-#' @seealso \href{https://developers.google.com/google-ads/api/fields/v10/campaign_query_builder}{Google Ads Query Builder}
+#' @seealso \href{https://developers.google.com/google-ads/api/fields/v23/campaign_query_builder}{Google Ads Query Builder}
 #'
 #' @return tibble with campaings dicrionary
 #' @export
@@ -37,8 +37,8 @@ gads_get_campaigns <- function(
                             'campaign.campaign_budget',
                             'campaign.bidding_strategy_type',
                             'campaign.dynamic_search_ads_setting.language_code',
-                            'campaign.start_date',
-                            'campaign.end_date',
+                            'campaign.start_date_time',
+                            'campaign.end_date_time',
                             'campaign.status',
                             'campaign.manual_cpm',
                             'campaign.manual_cpv',