+++ R/gads_get_report.R
@@ -1,7 +1,7 @@
 #' Get data from Google Ads API
 #'
-#' @param resource Report type, you can get list of all acessible resource using \code{\link{gads_get_metadata}}. For more information see \href{https://developers.google.com/google-ads/api/fields/v10/overview#list-of-all-resources}{link with list of all resources}
-#' @param fields character vector, list of report fields, all report has own fields list. You can get list of accesible resource fields using \code{\link{gads_get_fields}} for example \href{https://developers.google.com/google-ads/api/fields/v10/campaign}{see field list of campaign report}.
+#' @param resource Report type, you can get list of all acessible resource using \code{\link{gads_get_metadata}}. For more information see \href{https://developers.google.com/google-ads/api/fields/v23/overview#list-of-all-resources}{link with list of all resources}
+#' @param fields character vector, list of report fields, all report has own fields list. You can get list of accesible resource fields using \code{\link{gads_get_fields}} for example \href{https://developers.google.com/google-ads/api/fields/v23/campaign}{see field list of campaign report}.
 #' @param where Filter, for example you can filter campaigns by status \code{where = "campaign.status = 'ENABLED'"}.
 #' @param order_by Sorting, character vectors of fields and sorting directions, for example \code{order_by = c("campaign.name DESC", "metrics.clicks")}.
 #' @param limit Maximun rows in report
@@ -12,15 +12,15 @@
 #' @param customer_id Google Ads client customer id, supports a single account id: "xxx-xxx-xxxx" or a vector of ids from the same Google Ads MCC: c("xxx-xxx-xxxx", "xxx-xxx-xxxx")
 #' @param login_customer_id Google Ads manager customer id
 #' @param include_resource_name Get resource names fields in report
-#' @param gaql_query GAQL Query, you can make it in \code{\link{gads_get_metadata}}. For more information see \href{https://developers.google.com/google-ads/api/fields/v10/overview_query_builder}{Query Builder}. If you use gaql_query, you don't need set other query parameters like resource, fields, where, dates etc.
+#' @param gaql_query GAQL Query, you can make it in \code{\link{gads_get_metadata}}. For more information see \href{https://developers.google.com/google-ads/api/fields/v23/overview_query_builder}{Query Builder}. If you use gaql_query, you don't need set other query parameters like resource, fields, where, dates etc.
 #' @param verbose Console log output
 #' @param cl A cluster object created by \code{\link{makeCluster}}, or an integer to indicate number of child-processes (integer values are ignored on Windows) for parallel evaluations (see Details on performance).
 #'
 #' @return tibble with the Google Ads Data.
 #' @seealso
 #' \itemize{
-#'   \item \href{https://developers.google.com/google-ads/api/fields/v10/overview}{Oficial Google Ads API Reports documantation}
-#'   \item \href{https://developers.google.com/google-ads/api/fields/v10/overview_query_builder}{Google Ads Query Builder}
+#'   \item \href{https://developers.google.com/google-ads/api/fields/v23/overview}{Oficial Google Ads API Reports documantation}
+#'   \item \href{https://developers.google.com/google-ads/api/fields/v23/overview_query_builder}{Google Ads Query Builder}
 #' }
 #' @export
 #'